refactor: Header 전체적 코드 리팩터링, 모바일/태블릿/데스크탑으로 기존에 나누었던거 삭제후 Header으로 통합
